create types.rs
